collector/requirements.txt
psycopg2-binary==2.9.10
watchdog==5.0.3
